Use float64 instead of float32 for prices

+++ accounting/allocation.go
@@ -4,7 +4,7 @@ package accounting
 type Allocation struct {
 
 	// the amount being applied to the invoice
-	AppliedAmount float32 `json:"AppliedAmount,omitempty" xml:"AppliedAmount,omitempty"`
+	AppliedAmount float64 `json:"AppliedAmount,omitempty" xml:"AppliedAmount,omitempty"`
 
 	// the date the prepayment is applied YYYY-MM-DD (read-only). This will be the latter of the invoice date and the prepayment date.
 	Date string `json:"Date,omitempty" xml:"-"`
