+++ accounting/purchase_order.go
@@ -65,19 +65,19 @@ type PurchaseOrder struct {
 	PurchaseOrderID string `json:"PurchaseOrderID,omitempty" xml:"PurchaseOrderID,omitempty"`
 
 	// The currency rate for a multicurrency purchase order. As no rate can be specified, the XE.com day rate is used.
-	CurrencyRate float32 `json:"CurrencyRate,omitempty" xml:"CurrencyRate,omitempty"`
+	CurrencyRate float64 `json:"CurrencyRate,omitempty" xml:"CurrencyRate,omitempty"`
 
 	// Total of purchase order excluding taxes
-	SubTotal float32 `json:"SubTotal,omitempty" xml:"SubTotal,omitempty"`
+	SubTotal float64 `json:"SubTotal,omitempty" xml:"SubTotal,omitempty"`
 
 	// Total tax on purchase order
-	TotalTax float32 `json:"TotalTax,omitempty" xml:"TotalTax,omitempty"`
+	TotalTax float64 `json:"TotalTax,omitempty" xml:"TotalTax,omitempty"`
 
 	// Total of Purchase Order tax inclusive (i.e. SubTotal + TotalTax)
-	Total float32 `json:"Total,omitempty" xml:"Total,omitempty"`
+	Total float64 `json:"Total,omitempty" xml:"Total,omitempty"`
 
 	// Total of discounts applied on the purchase order line items
-	TotalDiscount float32 `json:"TotalDiscount,omitempty" xml:"TotalDiscount,omitempty"`
+	TotalDiscount float64 `json:"TotalDiscount,omitempty" xml:"TotalDiscount,omitempty"`
 
 	// boolean to indicate if a purchase order has an attachment
 	HasAttachments bool `json:"HasAttachments,omitempty" xml:"-"`
